# packages/types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requests": {
      "post": {
        "operationId": "RequestsController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-captcha-token",
            "required": false,
            "in": "header",
            "description": "Токен Яндекс SmartCaptcha — нужен при повторных отправках",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Ключ идемпотентности (8–64 символа)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRequestResponse"
                }
              }
            }
          },
          "428": {
            "description": "Требуется прохождение капчи"
          },
          "429": {
            "description": "Слишком много заявок с одного адреса"
          }
        },
        "summary": "Создать заявку «Перезвоните мне» (публичный, идемпотентен с Idempotency-Key)",
        "tags": [
          "requests"
        ]
      },
      "get": {
        "operationId": "RequestsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "new",
                "called",
                "taken",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Список заявок (по убыванию даты)",
        "tags": [
          "requests"
        ]
      }
    },
    "/requests/count-new": {
      "get": {
        "operationId": "RequestsController_countNew",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Количество новых заявок",
        "tags": [
          "requests"
        ]
      }
    },
    "/requests/{id}/status": {
      "patch": {
        "operationId": "RequestsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRequestStatusBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Изменить статус заявки и синхронизировать сообщение в Telegram",
        "tags": [
          "requests"
        ]
      }
    },
...
    "/bookings/slots": {
      "get": {
        "operationId": "BookingController_slots",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SlotView"
                  }
                }
              }
            }
          }
        },
        "summary": "Слоты дня со свободностью (публичный)",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings": {
      "post": {
        "operationId": "BookingController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-captcha-token",
            "required": false,
            "in": "header",
            "description": "Токен Яндекс SmartCaptcha — нужен при повторных отправках",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Ключ идемпотентности (8–64 символа)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBookingResponse"
                }
              }
            }
          },
          "409": {
            "description": "Слот занят или время прошло"
          },
          "428": {
            "description": "Требуется прохождение капчи"
          },
          "429": {
            "description": "Слишком много попыток с одного адреса"
          }
        },
        "summary": "Создать запись «Сам запишусь» (публичный, идемпотентен)",
        "tags": [
          "bookings"
        ]
      },
      "get": {
        "operationId": "BookingController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "booked",
                "confirmed",
                "taken",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Список записей (по убыванию даты создания)",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/day-overview": {
      "get": {
        "operationId": "BookingController_dayOverview",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {
              "example": "2026-11-12",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayOverviewResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обзор дня для настройки блокировок: слоты с занятостью",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/count-unconfirmed": {
      "get": {
        "operationId": "BookingController_countUnconfirmed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Количество записей, ждущих подтверждения",
        "tags": [
          "bookings"
        ]
      }
    },
    "/bookings/{id}/status": {
      "patch": {
        "operationId": "BookingController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingStatusBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Изменить статус записи и синхронизировать Telegram",
        "tags": [
          "bookings"
        ]
      }
    },
    "/booking-settings": {
      "get": {
        "operationId": "BookingController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Настройки записи",
        "tags": [
          "bookings"
        ]
      },
      "put": {
        "operationId": "BookingController_updateSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить настройки записи (частичное)",
        "tags": [
          "bookings"
        ]
      }
    },
...
  },
  "info": {
    "title": "STO API",
    "description": "API автосервиса: публичный сайт и админка",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "локальная разработка"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateRequestBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Иван",
            "minLength": 2,
            "maxLength": 80
          },
          "phone": {
            "type": "string",
            "example": "+375 29 123-45-67"
          },
          "comment": {
            "type": "string",
            "example": "Машину тянет вправо, звоните после 18:00",
            "maxLength": 500
          },
          "company": {
            "type": "string",
            "description": "Honeypot-поле: заполнено только ботами"
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
      "CreateRequestResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "duplicate": {
            "type": "boolean",
            "example": false,
            "description": "true — заявка уже существовала (повторный Idempotency-Key)"
          }
        },
        "required": [
          "id",
          "duplicate"
        ]
      },
      "UpdateRequestStatusBody": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "new",
              "called",
              "taken",
              "cancelled"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "SlotView": {
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "example": "14:00"
          },
          "available": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "time",
          "available"
        ]
      },
      "CreateBookingBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Иван",
            "minLength": 2,
            "maxLength": 80
          },
          "phone": {
            "type": "string",
            "example": "+375 29 123-45-67"
          },
          "company": {
            "type": "string",
            "description": "Honeypot-поле: заполнено только ботами"
          },
          "car": {
            "type": "string",
            "example": "Toyota Corolla 2015"
          },
          "services": {
            "example": [
              "Диагностика подвески",
              "Замена масла"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comment": {
            "type": "string",
            "example": "Стучит спереди на кочках",
            "maxLength": 500
          },
          "date": {
            "type": "string",
            "example": "2026-11-12",
            "description": "YYYY-MM-DD"
          },
          "time": {
            "type": "string",
            "example": "14:00",
            "description": "HH:MM"
          }
        },
        "required": [
          "name",
          "phone",
          "date",
          "time"
        ]
      },
      "CreateBookingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "duplicate": {
            "type": "boolean",
            "example": false,
            "description": "true — запись уже существовала (повторный Idempotency-Key)"
          }
        },
        "required": [
          "id",
          "duplicate"
        ]
      },
      "DayOverviewSlot": {
        "type": "object",
        "properties": {
          "time": {
            "type": "string",
            "example": "14:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "free",
              "booked",
              "blocked"
            ]
          },
          "past": {
            "type": "boolean",
            "example": false,
            "description": "true — слот уже прошёл"
          },
          "bookingNames": {
            "example": [
              "Иван"
            ],
            "description": "Имена клиентов с активной записью на слот",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "time",
          "status",
          "past",
          "bookingNames"
        ]
      },
      "DayOverviewResponse": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-11-12"
          },
          "enabled": {
            "type": "boolean",
            "example": true,
            "description": "false — день закрыт (выходной)"
          },
          "from": {
            "type": "string",
            "example": "09:00"
          },
          "to": {
            "type": "string",
            "example": "20:00"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOverviewSlot"
            }
          }
        },
        "required": [
          "date",
          "enabled",
          "slots"
        ]
      },
      "UpdateBookingStatusBody": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "booked",
              "confirmed",
              "taken",
              "cancelled"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}